cern

+++ src/pages/core/card/CardsStacksViews.tsx
@@ -1,11 +1,10 @@
 import { DndContext } from '@dnd-kit/core';
-import { User } from 'firebase/auth';
 import { deleteDoc, doc, DocumentData, getDoc, updateDoc } from "firebase/firestore";
 import { useState } from 'react';
 import { dbservice } from 'src/baseApi/serverbase';
+import useSelectors from 'src/hooks/useSelectors';
 import CardDroppable from './CardsDroppable';
 import CardsStacksViewsCollection from './CardsStacksViewsCollection';
-import useSelectors from 'src/hooks/useSelectors';
 import deleteMessage from './deleteMessage';
 
 const handleDelete = async ({
@@ -18,13 +17,18 @@ const handleDelete = async ({
   changeLongPressCard: (newValue: string) => void
 }) => {
   const data = doc(dbservice, `num/${id}`)
-  const messageId = data.id
-  await deleteDoc(doc(dbservice, `num/${id}`));
-  const userRef = doc(dbservice, `members/${profile?.uid}`)
-  const userSnap = await getDoc(userRef)
-  const newMessages = userSnap.data()?.createdCards.filter((element: string) => element !== id)
-  updateDoc(userRef, { createdCards: newMessages })
-  deleteMessage(messageId)
+  const snap = await getDoc(data)
+  if (snap.data().round === 1) {
+    const messageId = data.id
+    await deleteDoc(doc(dbservice, `num/${id}`));
+    const userRef = doc(dbservice, `members/${profile?.uid}`)
+    const userSnap = await getDoc(userRef)
+    const newMessages = userSnap.data()?.createdCards.filter((element: string) => element !== id)
+    updateDoc(userRef, { createdCards: newMessages })
+    deleteMessage(messageId)
+  } else {
+    alert('Processing cards cannot be deleted')
+  }
   changeLongPressCard('')
 }
 
@@ -33,7 +37,7 @@ const CardsStacksViews = ({
 }: {
   messages: DocumentData[]
 }) => {
-  const [longPressCard, setLongPressCard] = useState('')
+  const [longPressCard, setLongPressCard] = useState(undefined)
   const profile = useSelectors((state) => state.profile.value)
   const changeLongPressCard = (newValue: string) => setLongPressCard(newValue)
   return (
